code fix size_t to int

+++ 0x1E-search_algorithms/0-linear.c
@@ -9,11 +9,11 @@
 */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i = 0;
+	int i = 0;
 
-	while (*array && array && i < size)
+	while (*array && array && i < (int)size)
 	{
-		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
+		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 		i++;
